Atualizacao para a versao 1.3.0: Com suporte ao Primefaces mais atual e otmizacoes na utilizacao da interface

+++ sharkness/src/main/java/org/sharkness/jsf/support/ModelController.java
@@ -31,11 +31,11 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 	private Class<T> klassModel;
 
 	private LazyDataModel<T> dataModel;
-	
+
 	protected Logger getLogger() {
 		return LoggerFactory.getLogger();
 	}
-	
+
 	public void setModelService(Service modelService) {
 		this.modelService = modelService;
 	}
@@ -44,102 +44,102 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 		if (this.modelService == null) setModelService(getModelServiceFromContext());
 		return this.modelService;
 	}
-	
+
 	private Service getModelServiceFromContext() {
-		
+
 		try {
-			
+
 			getLogger().info("ModelController.getModelServiceFromContext: Loading...");
-			
+
 			Class<Service> klassService = ModelFactory.getClassModelService(this);
-			
+
 			Service service = ctx().getBean(klassService);
-			
+
 			getLogger().info("ModelController.getModelServiceFromContext: Loaded.");
 
 			return service;
-		
+
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.getModelServiceFromContext", e);
-			
+
 			return null;
-		
+
 		}
-	
+
 	}
 
 	public void init() throws Exception {
-		
+
 		getLogger().info("ModelController.init: Loading...");
-		
+
 		for (Field f : this.getClass().getDeclaredFields()) {
 			if (f.isAnnotationPresent(Autowired.class)) {
 				if (!f.isAccessible()) f.setAccessible(true);
 				f.set(this, ctx().getBean(f.getType()));
 			}
 		}
-		
+
 		getLogger().info("ModelController.init: Loaded.");
-		
+
 	}
-	
+
 	public ModelController() {
-		
+
 		try {
-			
+
 			getLogger().info(new StringBuilder("ModelController.constructor[")
 				.append(this.getClass().getSimpleName()).append("]: Loading...")
 			.toString());
-			
+
 			this.klassModel = ModelFactory.getClass(this);
-			
+
 			model = newInstanceModel();
-			
+
 			createLazyModel();
-			
+
 			init();
-			
+
 			getLogger().info(new StringBuilder("ModelController.constructor[")
 				.append(this.getClass().getSimpleName()).append("]: Loaded.")
 			.toString());
-		
+
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.constructor", e);
-		
+
 		}
-	
+
 	}
 
 	private void createLazyModel() {
-		
+
 		getLogger().info("ModelController.createLazyModel: Creating...");
-		
+
 		this.dataModel = new LazyDataModel<T>() {
-			
+
 			private List<T> list;
-			
+
 			@Override
 			public List<T> load(int first, int pageSize, String sortField, SortOrder sortOrder, Map<String, String> filters) {
-			
+
 				List<T> models = (List<T>) getModelService().getListPagination(first, pageSize, sortField, convertSortOrder(sortOrder), filters, getLocaleContext());
-				
+
 				this.setRowCount(getModelService().getSizePagination(filters, getLocaleContext()));
-				
+
 				this.setPageSize(pageSize);
-				
+
 				this.list = models;
-				
+
 				return models;
-			
+
 			}
-		
+
 			@Override
 			public Object getRowKey(T model) {
 				return model.getId();
 			}
-			
+
 			@Override
 			public T getRowData(String modelId) {
 
@@ -148,20 +148,20 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 						return obj;
 					}
 				}
-				
+
 				return null;
 			}
 
 		};
-	
+
 		getLogger().info("ModelController.createLazyModel: Created...");
 
 	}
 
 	public List<T> getListModel() throws Exception {
-	
+
 		try {
-			
+
 			getLogger().debug(new StringBuilder("ModelController.getListModel: ")
 				.append(getModelService().getObjClass().getSimpleName())
 			.toString());
@@ -169,11 +169,11 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 			return getModelService().list();
 
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.list", e);
-			
+
 			throw e;
-			
+
 		}
 
 	}
@@ -181,7 +181,7 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 	public LazyDataModel<T> getListDataModel() throws Exception {
 
 		getLogger().info("ModelController.getListDataModel: getting...");
-		
+
 		return dataModel;
 
 	}
@@ -189,23 +189,27 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 	public T getModel() {
 
 		getLogger().info("ModelController.getModel: getting...");
-		
+
 		if (model == null) model = newInstanceModel();
-		
+
 		return model;
-	
+
 	}
 
 	public void setModel(T model) {
 
 		getLogger().info("ModelController.setModel: setting...");
-		
+
 		this.model = model;
+
+	}
 	
+	public Boolean getEditMode() {
+		return getModel().getId() != null;
 	}
 
 	public void prepareAddModel(ActionEvent actionEvent) {
-		
+
 		getLogger().info("ModelController.prepareAddModel: preparing...");
 
 		model = newInstanceModel();
@@ -216,63 +220,69 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 	}
 
 	public String deleteModel() throws Exception {
+		
+		if (getEditMode()) {
+			
+			try {
 
-		try {
-			
-			getLogger().debug(new StringBuilder("ModelController.delete[")
-				.append(getModelService().getObjClass().getSimpleName()).append("](id=").append(getModel().getId()).append(")")
-			.toString());
+				getLogger().debug(new StringBuilder("ModelController.delete[")
+					.append(getModelService().getObjClass().getSimpleName()).append("](id=").append(getModel().getId()).append(")")
+				.toString());
 
-			getModelService().delete(getModel().getId());
-		
-		} catch (DataIntegrityViolationException e) {
-			
-			I18n i18n = I18n.get("global.delete.dataIntegrityViolationException");
-		
-			message(i18n);
-			
-			getLogger().error(i18n.toString(), e);
-		
-		} catch (Exception e) {
-			
-			I18n i18n = I18n.get("global.delete.exception").add(" ERR: ").add(e.getMessage());
-			
-			message(i18n);
-			
-			getLogger().error(i18n.toString(), e);
-		
+				getModelService().delete(getModel().getId());
+
+			} catch (DataIntegrityViolationException e) {
+
+				I18n i18n = I18n.get("global.delete.dataIntegrityViolationException");
+
+				message(i18n);
+
+				getLogger().error(i18n.toString(), e);
+
+			} catch (Exception e) {
+
+				I18n i18n = I18n.get("global.delete.exception").add(" ERR: ").add(e.getMessage());
+
+				message(i18n);
+
+				getLogger().error(i18n.toString(), e);
+
+			}
+
 		}
+
+		refresh();
 		
 		return "index";
-	
+
 	}
 
 	public boolean addModel(ActionEvent actionEvent) throws Exception {
 
 		try {
-			
+
 			getLogger().debug(new StringBuilder("ModelController.addModel[")
 				.append(getModelService().getObjClass().getSimpleName()).append("]: adding...")
 			.toString());
-	
+
 			getModelService().insert(modelProcessBeforeSave(model));
-		
+
 			return true;
 
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.editModel", e);
-			
+
 			throw e;
-			
+
 		}
 
 	}
 
 	public boolean editModel(ActionEvent actionEvent) throws Exception {
-		
+
 		try {
-			
+
 			getLogger().debug(new StringBuilder("ModelController.editModel[")
 				.append(getModelService().getObjClass().getSimpleName()).append("]: editing...")
 			.toString());
@@ -282,17 +292,17 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 			return true;
 
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.editModel", e);
-			
+
 			throw e;
-			
+
 		}
-	
+
 	}
 
 	public void saveModel(ActionEvent actionEvent) throws Exception {
-		
+
 		getLogger().debug(new StringBuilder("ModelController.saveModel[")
 			.append(getModelService().getObjClass().getSimpleName()).append("]: saving...")
 		.toString());
@@ -301,48 +311,48 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 
 			boolean saved = false;
 
-			if (getModel().getId() != null) {
-				
+			if (getEditMode()) {
+
 				saved = editModel(actionEvent);
-			
+
 			} else {
-				
+
 				saved = addModel(actionEvent);
-			
+
 			}
-			
+
 			getLogger().debug(new StringBuilder("ModelController.saveModel[")
 				.append(getModelService().getObjClass().getSimpleName()).append("]: saved = ").append(saved)
 			.toString());
 
 			if (saved) {
-				
+
 				I18n i18n = I18n.get("global.save.success");
-				
+
 				message(i18n);
-				
+
 				getLogger().debug(new StringBuilder("ModelController.saveModel[")
 					.append(getModelService().getObjClass().getSimpleName()).append("]: message = ").append(i18n.toString())
 				.toString());
 
 			}
-		
+
 		} catch (Exception e) {
-			
+
 			I18n i18n = I18n.get("global.save.failure");
-			
+
 			message(i18n);
-			
+
 			getLogger().error(i18n.toString(), e);
-		
+
 		}
-	
+
 	}
 
 	protected T modelProcessBeforeSave(T model) {
 
 		getLogger().debug("ModelController.modelProcessBeforeSave: Not implemented...");
-		
+
 		return model;
 
 	}
@@ -350,13 +360,13 @@ public abstract class ModelController<IdType extends Serializable, T extends Mod
 	private T newInstanceModel() {
 
 		try {
-			
+
 			getLogger().debug("ModelController.newInstanceModel: instantiating...");
 
 			return klassModel.newInstance();
 
 		} catch (Exception e) {
-			
+
 			getLogger().error("ModelController.newInstanceModel", e);
 
 			return null;
